[#11] Use optparse-applicative for parsing configuration (#15)
* [#11] Use optparse-applicative for parsing configuration

This introduces a `Context` type to contain
the settings for the program, and some basic
scaffolding to parse and consume this record.

* Move Context to Hintman.Cli

Allows us to more easily expand context with more
command line arguments as necessary. The Context
type can remain a newtype so long as it only contains
just a single boolean field.

* Make imports from Options.Applicative explicit

* [#11] Use optparse-applicative for parsing configuration

This introduces a `Context` type to contain
the settings for the program, and some basic
scaffolding to parse and consume this record.

* Move Context to Hintman.Cli

Allows us to more easily expand context with more
command line arguments as necessary. The Context
type can remain a newtype so long as it only contains
just a single boolean field.

* Make imports from Options.Applicative explicit

* Fix build errors because of usage of string

* Integrate Hintman.App into Hintman

`runHintman` now parses cli arguments, and
then starts a server. `Hintman.App` is removed.

* Remove Hintman.App from cabal file

Co-Authored-By: cronokirby <cronokirby@gmail.com>